Move mdspan traits to type_utils
Move the mdspan type checks from make_tensor.h to
type_utils.h. Keep the make_tensor function
unchanged.
