+++ include/matx/core/make_tensor.h
@@ -35,7 +35,6 @@
 
 #include <cstddef>
 #include <cstdint>
-#include <cuda/std/mdspan>
 #include <limits>
 #include <type_traits>
 
@@ -44,74 +43,10 @@
 #include "matx/core/tensor_desc.h"
 #include "matx/core/dlpack.h"
 #include "matx/core/log.h"
+#include "matx/core/type_utils.h"
 
-#if __has_include(<mdspan>)
-  #include <mdspan>
-#endif
+
 namespace matx {
-
-namespace detail {
-
-template <typename T>
-struct mdspan_traits {
-  static constexpr bool is_mdspan = false;
-};
-
-// If the type is a CUDA mdspan, check whether its layout and accessor
-// can be used with MatX
-template <typename ElementType,
-          typename Extents,
-          typename LayoutPolicy,
-          typename AccessorPolicy>
-struct mdspan_traits<
-    cuda::std::mdspan<
-        ElementType,
-        Extents,
-        LayoutPolicy,
-        AccessorPolicy>> {
-  static constexpr bool is_mdspan = true;
-
-  // Checks whether the mdspan layout is supported by MatX
-  static constexpr bool has_supported_layout =
-      std::is_same_v<LayoutPolicy, cuda::std::layout_right> ||
-      std::is_same_v<LayoutPolicy, cuda::std::layout_left> ||
-      std::is_same_v<LayoutPolicy, cuda::std::layout_stride>;
-
-  // Rejects custom accessor functions
-  static constexpr bool has_default_accessor =
-      std::is_same_v<
-          AccessorPolicy,
-          cuda::std::default_accessor<ElementType>>;
-};
-
-// Type checker for the official C++23 std::mdspan
-#if defined(__cpp_lib_mdspan) && (__cpp_lib_mdspan >= 202207L)
-template <typename ElementType,
-          typename Extents,
-          typename LayoutPolicy,
-          typename AccessorPolicy>
-struct mdspan_traits<
-    std::mdspan<
-        ElementType,
-        Extents,
-        LayoutPolicy,
-        AccessorPolicy>> {
-  static constexpr bool is_mdspan = true;
-
-  static constexpr bool has_supported_layout =
-      std::is_same_v<LayoutPolicy, std::layout_right> ||
-      std::is_same_v<LayoutPolicy, std::layout_left> ||
-      std::is_same_v<LayoutPolicy, std::layout_stride>;
-
-  static constexpr bool has_default_accessor =
-      std::is_same_v<
-          AccessorPolicy,
-          std::default_accessor<ElementType>>;
-};
-#endif
-
-} 
-
 
 /**
  * Create a tensor with a C array for the shape using implicitly-allocated memory
